grizzly_data_lineage/requirements.txt
Flask~=2.1.2
mo-sql-parsing
PyYAML~=6.0
Jinja2~=3.1.2
python-dotenv~=0.20.0
gunicorn~=20.1.0
google-cloud-bigquery~=3.2.0
google-auth~=2.9.0
sql-formatter